+++ CHANGELOG.md
@@ -5,6 +5,13 @@
     - `RegExp.escape`
   - Moved to stable ES, [February 2025 TC39 meeting](https://github.com/tc39/proposals/commit/b81fa9bccf4b51f33de0cbe797976a84d05d4b76)
   - Added `es.` namespace module, `/es/` and `/stable/` namespaces entries
+- [`Float16` proposal](https://github.com/tc39/proposal-regex-escaping):
+  - Built-ins:
+    - `Math.f16round`
+    - `DataView.prototype.getFloat16`
+    - `DataView.prototype.setFloat16`
+  - Moved to stable ES, February 2025 TC39 meeting
+  - Added `es.` namespace modules, `/es/` and `/stable/` namespaces entries
 - Compat data improvements:
   - [`DisposableStack`, `AsyncDisposableStack`, `SuppressedError` and `Iterator.prototype[@@dispose]`](https://github.com/tc39/proposal-explicit-resource-management) marked as [shipped from V8 ~ Chromium 134](https://issues.chromium.org/issues/42203506#comment24)
   - [`Error.isError`](https://github.com/tc39/proposal-is-error) added and marked as supported from V8 ~ Chromium 134
