+++ packages/core-js-pure/override/modules/es.data-view.set-float16.js
@@ -1,0 +1,1 @@
+// empty
