+++ packages/core-js/actual/data-view/get-float16.js
@@ -1,2 +1,5 @@
 'use strict';
+var parent = require('../../stable/data-view/get-float16');
 require('../../modules/esnext.data-view.get-float16');
+
+module.exports = parent;
