+++ packages/core-js/actual/data-view/set-float16.js
@@ -1,2 +1,5 @@
 'use strict';
+var parent = require('../../stable/data-view/set-float16');
 require('../../modules/esnext.data-view.set-float16');
+
+module.exports = parent;
