+++ packages/core-js/es/data-view/get-float16.js
@@ -1,0 +1,2 @@
+'use strict';
+require('../../modules/es.data-view.get-float16');
