+++ packages/core-js/es/data-view/index.js
@@ -2,6 +2,8 @@
 require('../../modules/es.array-buffer.constructor');
 require('../../modules/es.array-buffer.slice');
 require('../../modules/es.data-view');
+require('../../modules/es.data-view.get-float16');
+require('../../modules/es.data-view.set-float16');
 require('../../modules/es.object.to-string');
 var path = require('../../internals/path');
