+++ packages/core-js/es/data-view/set-float16.js
@@ -1,0 +1,2 @@
+'use strict';
+require('../../modules/es.data-view.set-float16');
