+++ packages/core-js/es/math/f16round.js
@@ -1,0 +1,5 @@
+'use strict';
+require('../../modules/es.math.f16round');
+var path = require('../../internals/path');
+
+module.exports = path.Math.f16round;
