+++ packages/core-js/stable/data-view/get-float16.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/data-view/get-float16');
+
+module.exports = parent;
