+++ packages/core-js/stable/math/f16round.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/math/f16round');
+
+module.exports = parent;
