+++ packages/core-js/stage/3.js
@@ -5,7 +5,6 @@ require('../proposals/array-buffer-base64');
 require('../proposals/array-from-async-stage-2');
 require('../proposals/decorator-metadata-v2');
 require('../proposals/explicit-resource-management');
-require('../proposals/float16');
 require('../proposals/is-error');
 require('../proposals/json-parse-with-source');
 require('../proposals/math-sum');
