+++ tests/compat-tools/compat.mjs
@@ -16,6 +16,7 @@ deepEqual(compat({
     'es.array.at',
     'es.math.clz32',
     'es.math.expm1',
+    'es.math.f16round',
     'es.math.to-string-tag',
     'es.reflect.apply',
     'es.reflect.construct',
@@ -35,6 +36,7 @@ deepEqual(compat({
     'es.array.at': { firefox: '27' },
     'es.math.clz32': { firefox: '27' },
     'es.math.expm1': { firefox: '27' },
+    'es.math.f16round': { firefox: '27' },
     'es.math.to-string-tag': { firefox: '27' },
     'es.reflect.apply': { firefox: '27' },
     'es.reflect.construct': { firefox: '27' },
@@ -107,6 +109,7 @@ deepEqual(compat({
     'es.math.acosh',
     'es.math.clz32',
     'es.math.expm1',
+    'es.math.f16round',
     'es.math.hypot',
     'es.math.to-string-tag',
   ],
@@ -114,6 +117,7 @@ deepEqual(compat({
     'es.math.acosh': { chrome: '40' },
     'es.math.clz32': { firefox: '27' },
     'es.math.expm1': { firefox: '27' },
+    'es.math.f16round': { chrome: '40', firefox: '27' },
     'es.math.hypot': { chrome: '40' },
     'es.math.to-string-tag': { chrome: '40', firefox: '27' },
   },
